profiles/powersave: use conservative governor
Because of it's slow rampup, conservative avoids the very high frequency
p-states that waste energy in V^2 loss, except in cases of sustained
high CPU usage, which would typically correspond to "stopwatch"
workloads where the user is actively waiting on the computer.

This causes jank in interactive tasks, but, on HWP-capable CPUs, this
profile sets EPP all the way to power, which locks frequency to most
efficient (equivalent to AMD "lowest nonlinear perf"), so the tradeoff
conservative gives is consistent with the design intent.